pyproject.toml
[tool.poetry]
name = "newscollection"
version = "0.1.0"
description = ""
authors = ["Chuan <kirk336neo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
loguru = "^0.7.2"
motor = "^3.6.0"
aiohttp = "^3.10.10"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
ruff = "^0.6.9"
pytest = "^8.3.3"
testcontainers = "^4.8.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]


[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
pythonpath = "."
addopts = [
    "--import-mode=importlib",
]

